requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
beautifulsoup4==4.14.3
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.1.8
distro==1.9.0
dnspython==2.7.0
exceptiongroup==1.3.1
fastapi==0.128.8
filelock==3.19.1
fsspec==2025.10.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
jiter==0.14.0
llvmlite==0.43.0
MarkupSafe==3.0.3
more-itertools==10.8.0
motor==3.7.1
mpmath==1.3.0
networkx==3.2.1
numba==0.60.0
numpy==2.0.2
openai==2.32.0
openai-whisper==20250625
pydantic==2.13.3
pydantic_core==2.46.3
pymongo==4.17.0
python-multipart==0.0.20
regex==2026.1.15
requests==2.32.5
sniffio==1.3.1
soupsieve==2.8.3
starlette==0.49.3
sympy==1.14.0
tiktoken==0.12.0
torch==2.8.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.39.0
yt-dlp==2025.10.14
python-dotenv==1.0.1
Pillow
yt-dlp